requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
ollama>=0.1.7
pymupdf>=1.23.0
pdfplumber>=0.10.1
python-docx>=0.8.11
chromadb>=0.4.0
sentence-transformers>=2.2.0
guardrails-ai>=0.4.0
pydantic>=2.0.0
requests>=2.31.0
Pillow>=9.0.0
pandas>=1.5.0